+++ src/components/chat/ChatBubble.tsx
